buses_frontend/requirements.txt
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
Flask==3.1.0
fpdf==1.7.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jwt==1.3.1
MarkupSafe==3.0.2
pycparser==2.22
python-dotenv==1.0.1
qrcode==8.0
requests==2.32.3
urllib3==2.3.0
Werkzeug==3.1.3
